==> Codice_calibrazione/02_Create_Classe_e_causale_residua.ipynb <==
import pandas as pd
import os

input_dir = "00_Input"
output_dir = "01_output"

### Create SAE ATECO normalized for SAE_ATECO_NEW

# Load each file into its own DataFrame
df_classe= pd.read_excel(os.path.join(input_dir, "SAE_ATECO_NEW.xlsx"))
df_classe.columns

df_classe["ATECO_CODE"] = df_classe["ATECO"].str.replace("\xa0", " ", regex=True).str[:3]
df_classe[['ATECO','ATECO_CODE']].drop_duplicates()

df_classe["SAE_CODE"] = df_classe["SAE"].str.replace("\xa0", " ", regex=True).str[:3]
df_classe[['SAE','SAE_CODE']].drop_duplicates()


Creation of normalized SAE ATECO and testing it

df_classe["SAE_ATECO_CODE"] = (df_classe["SAE_CODE"].astype(str) + df_classe["ATECO_CODE"].astype(str))

print(set(df_classe["ATECO_CODE"].astype(str).str.len()))
print(set(df_classe["SAE_CODE"].astype(str).str.len()))
print(set(df_classe["SAE_ATECO_CODE"].astype(str).str.len()))

test = df_classe[['SAE_CODE','ATECO_CODE', 'SAE_ATECO_CODE','sae ateco']].drop_duplicates()
test = test[test['ATECO_CODE'] == '071']
test

## Create the SAE ATECO normalized for the periemter

df_perimeter = pd.read_excel(os.path.join(output_dir, "NETECH_tutti_scen_stato_v3.xlsx"))
df_perimeter.shape

df_perimeter["ATECO_CODE"] = df_perimeter["D22 ATECO"].str.replace("\xa0", " ", regex=True).str[:3]
df_perimeter["ATECO_CODE"] = df_perimeter["ATECO_CODE"].fillna("999")
df_perimeter[['D22 ATECO','ATECO_CODE']].drop_duplicates()

test = df_perimeter[df_perimeter['DESCRIZIONE'] == "Versamento giornaliero di importo elevato per la causale ed il soggetto"]
test[['D22 ATECO','ATECO_CODE']].drop_duplicates()

df_perimeter["SAE_CODE"] = df_perimeter["D21 SAE"].str.replace("\xa0", " ", regex=True).str[:3]
df_perimeter[['D21 SAE','SAE_CODE']].drop_duplicates()


df_perimeter["SAE_ATECO_CODE"] = (df_perimeter["SAE_CODE"].astype(str) + df_perimeter["ATECO_CODE"].astype(str))

lengths = df_perimeter["SAE_ATECO_CODE"].astype(str).str.len()
print(set(lengths))
df_perimeter[['SAE_CODE', 'ATECO_CODE', 'SAE_ATECO_CODE']].drop_duplicates()


test = df_perimeter[['SAE_CODE', 'ATECO_CODE', 'SAE_ATECO_CODE']]
test[test['SAE_CODE'] == '284']

Check that every SAE in Netech data is present in SAE ATECO excel

len(set(df_perimeter["SAE_CODE"])-set(df_classe["SAE_CODE"]))

Check that every ATECO in Netech data is present in SAE ATECO excel

ateco_difference = set(df_perimeter["ATECO_CODE"])-set(df_classe["ATECO_CODE"])
print(len(ateco_difference))

df_joined = df_perimeter.merge(
    df_classe[["SAE_ATECO_CODE", "RISCHIO CONTANTE"]].drop_duplicates(),
    how="left",
    on="SAE_ATECO_CODE")

print(set(df_joined['RISCHIO CONTANTE']))

print(df_perimeter.shape)
print(df_joined.shape)

Mapping the residual classes

mask = df_joined["RISCHIO CONTANTE"].isna()
df_joined.loc[mask, "ATECO_CODE"] = "999"
df_joined.loc[mask, "RES_ATECO"] = '1'

print(df_joined[mask].shape)

df_joined = df_joined.rename(columns={"RISCHIO CONTANTE": 'Classe_old'})

df_joined.loc[mask, "SAE_ATECO_CODE"] = (df_joined.loc[mask, "SAE_CODE"].astype(str) + df_joined.loc[mask, "ATECO_CODE"].astype(str))

df_joined = df_joined.merge(
    df_classe[["SAE_ATECO_CODE", "RISCHIO CONTANTE"]].drop_duplicates(),
    how="left",
    on="SAE_ATECO_CODE")

df_joined = df_joined.rename(columns={"RISCHIO CONTANTE": 'Classe'})
print(df_joined.shape)

mask_ = df_joined["RES_ATECO"] == '1'
df_joined.loc[~mask_,'Classe_old'].value_counts()

mask_ = df_joined["RES_ATECO"] == '1'
df_joined.loc[~mask_,'Classe'].value_counts()

Replace causale sportello and VV with SPECIAL

Special_Causale = ['451', '452', '14M', '56', 'ES', 'Y11', 'Y12', 'Y13', 'AFI']
df_joined["CAUSALE ANALITICA"] = df_joined["CAUSALE ANALITICA"].replace(Special_Causale, "SPECIAL")
df_joined.shape

output_file = os.path.join(output_dir, "NETECH_CAMPIONE_CALIB.xlsx")
# Export DataFrame to Excel
df_joined.to_excel(output_file, index=False)
